requirements.txt
fastapi
uvicorn
pydantic
assemblyai
google-genai
python-multipart
python-dotenv
pydantic_settings